added client-outlet check condition in open-orders endpoint
